Enhanced access token handling with Flask-JWT-Extended.

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 
@@ -16,7 +16,7 @@ class Item(Resource):
                         help="Store ID cannot be empty"
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         try:
             item = ItemModel.find_by_name(name)
@@ -41,7 +41,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'
+                    }
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -70,7 +75,14 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
-        # alternatively, {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
 
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'item': [item['name'] for item in items],
+            'message': 'More data available if you log in'
+        }, 200
